feat: Rework api to allow comparing of JsonNumber
JsonNumber was originally Option (Int), Option (Float), Option (String)
This will cause problems in the future and this is going to be a
breaking change. Having the string is useful to pick out the original
value in the JSON string but this causes issues comparing a basic
JsonNumber with one constructed from parsing as the original string
could be significantly different. Also Floats with no decimal element
will now be converted to Ints to make comparisons accurate